+++ packages/cloud-cognitive/src/components/CreateTearsheet/CreateTearsheet.js
@@ -113,7 +113,9 @@ export let CreateTearsheet = forwardRef(
             setIsSubmitting(false);
             console.warn(`${componentName} onNext error: ${error}`);
           }
-        } else continueToNextStep();
+        } else {
+          continueToNextStep();
+        }
       };
       const handleSubmit = async () => {
         setIsSubmitting(true);
@@ -127,19 +129,22 @@ export let CreateTearsheet = forwardRef(
             setIsSubmitting(false);
             console.warn(`${componentName} onNext error: ${error}`);
           }
-        } else await handleOnRequestSubmit();
+        } else {
+          await handleOnRequestSubmit();
+        }
       };
       if (getTearsheetSteps()?.length) {
         const createSteps = getTearsheetSteps();
         const total = createSteps.length;
         const buttons = [];
-        if (total > 1)
+        if (total > 1) {
           buttons.push({
             label: backButtonText,
             onClick: () => setCurrentStep((prev) => prev - 1),
             kind: 'secondary',
             disabled: currentStep === 1,
           });
+        }
         buttons.push({
           label: cancelButtonText,
           onClick: onUnmount,
@@ -222,26 +227,42 @@ export let CreateTearsheet = forwardRef(
         <>
           {' '}
           {childrenArray.map((child, stepIndex) => {
-            if (!isTearsheetStep(child)) return child;
+            if (!isTearsheetStep(child)) {
+              return child;
+            }
             step++;
-            return React.cloneElement(child, {
-              className: cx(child.props.className, {
-                [`${blockClass}__step--hidden-section`]: currentStep !== step,
-                [`${blockClass}__step--visible-section`]: currentStep === step,
-              }),
-              key: `key_${stepIndex}`,
-            });
+            return React.cloneElement(
+              child,
+              {
+                className: cx(child.props.className, {
+                  [`${blockClass}__step--hidden-section`]: currentStep !== step,
+                  [`${blockClass}__step--visible-section`]:
+                    currentStep === step,
+                  [`${blockClass}__step--first-panel-step`]:
+                    !previousState?.open &&
+                    open &&
+                    previousState?.currentStep === 0 &&
+                    stepIndex === 0,
+                }),
+                key: `key_${stepIndex}`,
+              },
+              <>
+                <p className={`${blockClass}__step--heading`}>
+                  {renderStepTitle(stepIndex)}
+                </p>
+                {child.props.children}
+              </>
+            );
           })}
         </>
       );
     };
 
     // renders the individual step title
-    const renderStepTitle = () => {
+    const renderStepTitle = (stepIndex) => {
       const tearsheetSteps = getTearsheetSteps();
       const stepTitle =
-        (tearsheetSteps && tearsheetSteps[currentStep - 1]?.props.title) ||
-        null;
+        (tearsheetSteps && tearsheetSteps[stepIndex]?.props.title) || null;
       return stepTitle;
     };
 
@@ -324,7 +345,6 @@ export let CreateTearsheet = forwardRef(
         <div
           className={`${blockClass}__multi-step-panel-content`}
           onBlur={handleBlur}>
-          <p className={`${blockClass}__step--heading`}>{renderStepTitle()}</p>
           {renderChildren(children)}
         </div>
       </TearsheetShell>