+++ packages/cloud-cognitive/src/components/TagSet/TagSetOverflow.js
@@ -14,7 +14,7 @@ import { Link, Tag, Tooltip } from 'carbon-components-react';
 
 import { pkg } from '../../settings';
 const componentName = 'TagSetOverflow';
-const blockClass = `${pkg.prefix}--tag-set`;
+const blockClass = `${pkg.prefix}--tag-set-overflow`;
 
 export const TagSetOverflow = React.forwardRef(
   (
@@ -25,13 +25,12 @@ export const TagSetOverflow = React.forwardRef(
       overflowAlign,
       overflowDirection,
       showAllTagsLabel,
+      allTagsModalSearchThreshold,
     },
     ref
   ) => {
     const [tipOpen, setTipOpen] = useState(false);
     const overflowTagContent = useRef(null);
-    const localRef = useRef(null);
-    const overflowRef = ref || localRef;
 
     const handleChange = (ev, { open }) => {
       setTipOpen(open);
@@ -47,10 +46,10 @@ export const TagSetOverflow = React.forwardRef(
     return (
       <span
         aria-hidden={overflowTags.length === 0}
-        className={cx(`${blockClass}__overflow`, {
-          [`${blockClass}__overflow--hidden`]: overflowTags.length === 0,
+        className={cx(`${blockClass}`, {
+          [`${blockClass}--hidden`]: overflowTags.length === 0,
         })}
-        ref={overflowRef}>
+        ref={ref}>
         <Tooltip
           align={overflowAlign}
           className={cx(className, `${blockClass}__tooltip`)}
@@ -59,25 +58,26 @@ export const TagSetOverflow = React.forwardRef(
           open={tipOpen}
           triggerText={<Tag>+{overflowTags.length}</Tag>}
           showIcon={false}>
-          <div
-            ref={overflowTagContent}
-            className={`${blockClass}__overflow-content`}>
-            <ul className={`${blockClass}__overflow-tag-list`}>
+          <div ref={overflowTagContent} className={`${blockClass}__content`}>
+            <ul className={`${blockClass}__tag-list`}>
               {overflowTags
-                .filter((_, index) => index < 10)
+                .filter((_, index) =>
+                  overflowTags.length > allTagsModalSearchThreshold
+                    ? index < allTagsModalSearchThreshold
+                    : index <= allTagsModalSearchThreshold
+                )
                 .map((tag, index) => (
-                  <li
-                    className={`${blockClass}__overflow-tag-item`}
-                    key={index}>
+                  <li className={`${blockClass}__tag-item`} key={index}>
                     {React.cloneElement(tag, { filter: false })}
                   </li>
                 ))}
             </ul>
-            {overflowTags.length >= 10 && (
+            {overflowTags.length > allTagsModalSearchThreshold && (
               <Link
                 className={`${blockClass}__show-all-tags-link`}
                 href=""
-                onClick={handleShowAllTagsClick}>
+                onClick={handleShowAllTagsClick}
+                role="button">
                 {showAllTagsLabel}
               </Link>
             )}
@@ -91,6 +91,10 @@ export const TagSetOverflow = React.forwardRef(
 TagSetOverflow.displayName = componentName;
 
 TagSetOverflow.propTypes = {
+  /**
+   * count of overflowTags over which a modal is offered
+   */
+  allTagsModalSearchThreshold: PropTypes.number,
   /**
    * className
    */
@@ -118,7 +122,7 @@ TagSetOverflow.propTypes = {
 };
 
 TagSetOverflow.defaultProps = {
+  allTagsModalSearchThreshold: 10,
   overflowAlign: 'center',
   overflowDirection: 'bottom',
-  showAllTagsLabel: 'View all tags',
 };