+++ packages/cloud-cognitive/src/components/Tearsheet/TearsheetNarrow.js
@@ -11,13 +11,18 @@ import React from 'react';
 // Other standard imports.
 import PropTypes from 'prop-types';
 import { pkg } from '../../settings';
-import { allPropTypes, prepareProps } from '../../global/js/utils/props-helper';
+import {
+  allPropTypes,
+  deprecateProp,
+  prepareProps,
+} from '../../global/js/utils/props-helper';
 
 // Carbon and package components we use.
 import { Button } from 'carbon-components-react';
 import { ActionSet } from '../ActionSet';
 
 import {
+  tearsheetHasCloseIcon,
   TearsheetShell,
   tearsheetShellWideProps as blocked,
 } from './TearsheetShell';
@@ -86,8 +91,13 @@ TearsheetNarrow.propTypes = {
 
   /**
    * The accessibility title for the close icon (if shown).
+   *
+   * **Note:** This prop is only required if a close icon is shown, i.e. if
+   * there are a no navigation actions and/or hasCloseIcon is true.
    */
-  closeIconDescription: PropTypes.string,
+  closeIconDescription: PropTypes.string.isRequired.if(
+    ({ actions, hasCloseIcon }) => tearsheetHasCloseIcon(actions, hasCloseIcon)
+  ),
 
   /**
    * A description of the flow, displayed in the header area of the tearsheet.
@@ -126,7 +136,10 @@ TearsheetNarrow.propTypes = {
    * provided, which can be cancelled by returning 'false') if the user clicks
    * outside it.
    */
-  preventCloseOnClickOutside: PropTypes.bool,
+  preventCloseOnClickOutside: deprecateProp(
+    PropTypes.bool,
+    'The tearsheet will close automatically if the user clicks outside it if and only if the tearsheet is passive (no navigation actions)'
+  ),
 
   /**
    * The main title of the tearsheet, displayed in the header area.
@@ -148,7 +161,5 @@ TearsheetNarrow.propTypes = {
 // 'undefined' values reasonably. Default values should be provided when the
 // component needs to make a choice or assumption when a prop is not supplied.
 TearsheetNarrow.defaultProps = {
-  closeIconDescription: 'Close',
-  hasCloseIcon: false,
   verticalPosition: 'normal',
 };